requirements.txt
requests
beautifulsoup4
flask
pillow
gunicorn
